// src/Models/brewer.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 07 Apr 2022 08:01:06 GMT'
 */
export type Brewer = IContentItem<{
  productName: Elements.TextElement;
  metadataOgDescription: Elements.TextElement;
  metadataMetaTitle: Elements.TextElement;
  longDescription: Elements.RichTextElement;
  metadataOgTitle: Elements.TextElement;
  metadataMetaDescription: Elements.TextElement;
  metadataTwitterSite: Elements.TextElement;
  price: Elements.NumberElement;
  manufacturer: Elements.TaxonomyElement;
  metadataTwitterImage: Elements.AssetsElement;
  metadataTwitterCreator: Elements.TextElement;
  urlPattern: Elements.UrlSlugElement;
  sitemap: Elements.TaxonomyElement;
  shortDescription: Elements.RichTextElement;
  productStatus: Elements.TaxonomyElement;
  metadataTwitterTitle: Elements.TextElement;
  metadataTwitterDescription: Elements.TextElement;
  metadataOgImage: Elements.AssetsElement;
  image: Elements.AssetsElement;
}>;